# expense-management-api/storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SpendWise API",
    "version": "1.0.0",
    "description": "API for managing expenses, budgets, balances and group contexts",
    "contact": {
      "email": "support@spendwise.com"
    }
  },
  "servers": [
    {
      "url": "http://expense-management-api.test//api",
      "description": "Local API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token in format: Bearer <token>"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": {"type": "string", "example": "Rukaiya"},
                  "email": {"type": "string", "format": "email", "example": "rukaiya@email.com"},
                  "password": {"type": "string", "example": "12345678"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"},
                    "user": {"type": "object"},
                    "token": {"type": "string"},
                    "token_type": {"type": "string"},
                    "expires_in": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login user and get JWT token",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {"type": "string", "format": "email", "example": "rukaiya@email.com"},
                  "password": {"type": "string", "example": "12345678"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {"type": "object"},
                    "token": {"type": "string"},
                    "token_type": {"type": "string"},
                    "expires_in": {"type": "integer"}
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["Password Reset"],
        "summary": "Request password reset email",
        "operationId": "forgotPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": {"type": "string", "format": "email", "example": "user@email.com"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset link sent"
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["Password Reset"],
        "summary": "Reset password with token",
        "operationId": "resetPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "password_confirmation", "token"],
                "properties": {
                  "email": {"type": "string", "format": "email"},
                  "password": {"type": "string"},
                  "password_confirmation": {"type": "string"},
                  "token": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully"
          }
        }
      }
    },
...
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout user",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Logged out successfully"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get authenticated user",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "User data"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh JWT token",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Token refreshed"
          }
        }
      }
    },
    "/auth/profile": {
      "patch": {
        "tags": ["Profile"],
        "summary": "Update profile",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {"type": "string"},
                  "avatar_url": {"type": "string", "format": "uri"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated"
          }
        }
      }
    },
...
    "/balances/summary": {
      "get": {
        "tags": ["Balances"],
        "summary": "Get balance summary (who owes whom)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "context_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {
            "description": "Balance summary"
          }
        }
      }
    },
    "/balances/settlements": {
      "post": {
        "tags": ["Balances"],
        "summary": "Record a settlement",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["context_id", "amount", "settled_by"],
                "properties": {
                  "context_id": {"type": "string", "format": "uuid"},
                  "amount": {"type": "number"},
                  "settled_by": {"type": "string", "format": "uuid"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Settlement recorded"
          }
        }
      },
      "get": {
        "tags": ["Balances"],
        "summary": "Get settlement history",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "context_id", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid"}},
          {"name": "per_page", "in": "query", "schema": {"type": "integer", "default": 20}}
        ],
        "responses": {
          "200": {
            "description": "Settlement history"
          }
        }
      }
    },
...
  }
}